Add files via upload
The NLTK (Natural Language Toolkit) and GUI (Graphical User Interface)-based SMS spam filtering methodology is a method for automatically identifying and removing undesired spam messages from a collection of SMS texts. NLTK, a potent Python toolkit for natural language processing, is used in this method to analyze message content and categories it as either spam or valid.